+++ CHANGELOG.md
@@ -5,14 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
 ### Added
-
 - Added Clippy to CI
 - Add `flip-link` to support examples using it
+- Add example for an IWDG on an STM32F4 with RTIC v1.0
+
+### Changed
 - Updated v0.6-alpha.5 project to RTIC v1.0
 
 ### Fixed
-
 - Use rust-cache instead of native GHA cache
 - Update stm32l4 to use heapless 0.6
 - Fix cargo fmt to actually fail CI